requirements.txt
flask
flask-cors
streamlit
google-generativeai
pdfplumber
pytesseract
Pillow
reportlab
joblib
scikit-learn
requests
